Refactor to get rid of async_trait
We now expose a channel between the search provider and the outer
application, and the outer application is expected to read requests of
that channel and respond accordingly.

This allows us to replace async_trait and have more freedom in the
implementation of search requests without being bound to the Send+Sync
requirements of zbus' dbus_interface.